+++ liter/__init__.py
@@ -1,5 +1,5 @@
 REPR_INDENT = 2
 
-from . import engine, events, exception, stub, utils, writer
+from . import engine, exception, stub, utils, writer
 
 __version__ = "0.1.8"
