+++ liter/engine/__init__.py
@@ -1,4 +1,5 @@
 from .. import REPR_INDENT
 from .base import *
 from .buffer import *
+from .events import *
 from .factory import *
